Adjust the provisioning of servers in playbooks
Move out the VM provisioning from the virthost setup so that you don't
have to continually run the virthost setup just to spin up a set of VMs.

Move the reading of the generated root SSH key into the vm-spinup
role.

Also, I renamed some things so that we use some of the same naming
conventions as APB just to start getting used to that formatting.