Rework openshift bootstrap role for baremetal
While working through the bootstrap role for baremetal it became
obvious that there could be a bit of tweaking to make the baremetal
support a little easier to consume.

The primary change this adds is an ability to skip over the thinpool
provisioning, because that could fairly easily be done either in the
qcow2 image ahead of time (bifrost example) or in the kickstart file
(cobbler example).

Also, instead of pushing the TLD of .example.local in the
configuration templates, that's been made a default value that you can
more easily override.

Built a default variable for the /dev/mapper/{{thinpool_name}} as well
which will make things more flexible if you're not naming your volume
group and thinpool name exactly as the static name.